requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.15.2,<3.16
psycopg2>=2.9.10,<2.10
python-dateutil>=2.9.0,<2.10
